Built-in scaling of the input features
Built-in scaling of the input features and a jupyter notebook show its performance and usage.
